Just added memo
I used an Overleaf template to construct this file, which has dependecy
on a .cls tex definition file. This hinders to file from being compiled
correctly by other tex editors such as Texworks. I inted on printing and
submitting the pdf myself due to this but let me know if this causes
problems with anyone and i will change the file.